Fix on registration. Problem with json format in svelte. Decoupling some classes
