+++ TheDailyRoutine/TheDailyRoutine/Models/ErrorResponse.cs
@@ -1,0 +1,7 @@
+﻿namespace TheDailyRoutine.Models
+{
+    public class ErrorResponse
+    {
+        public string Error { get; set; }
+    }
+}
