+++ TheDailyRoutine/TheDailyRoutine/Models/LoginRequest.cs
@@ -1,0 +1,8 @@
+﻿namespace TheDailyRoutine.Models
+{
+    public class LoginRequest
+    {
+        public string Identifier { get; set; }
+        public string Password { get; set; }
+    }
+}
